split png code into separate files, and use try-catch to avoid pragma disable
